+++ changelog.md
@@ -1,5 +1,6 @@
 # v1.0.1
 - Added clear search button
+- Added full Geode theme compatibility
 - Set default value for setting *`Clear Search on Refresh`* to `false`
 
 # v1.0.0